cloudsql/read-replica-autoscaling/requirements.txt
google-cloud-sql==1.12.0
google-cloud-monitoring==2.15.1
functions-framework==3.4.0